refactor(cosmo-pd101): split StepEnvelopeEditor into smaller modules
